Move js code to main
